==> class_exercises/03.rmd <==
---
title: "Git to BASH"
author: "JR"
date: "9/26/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Our first objective today will be to sync to the class github. We will
go over a few needed things to establish git connections:

### Sign up for github -- if you already have nothing to do.


## FIRST STEP: set up ssh connection to github (one time only)

Navigate to your home directory on fiji ~/IDENTIKEY

```{bash}

cd ~
ls -lah

```

Note that there is a .ssh folder that you can place these unique keys to let
servers communicate securely. We can check to see if a key already exists.
If it doesn't we will generate one with the with ssh-keygen command.

```{bash}
cd .ssh
ls -lah

# If it exists already, you can check out the public key using cat
cat id_rsa.pub

```


# Making a new RSA keygen if you don't already have one
```{bash}

ssh-keygen -t rsa

# Then cat the contents of the public key file (id_rsa.pub) and copy to clipboard.

```


Now let's got to your github website. You will need to sign into your github account first.

On the top right there is a drop down
menu. 

Click on "Settings". Now you will see on the left side there is a list of
options. Click on "SSH and GPG Keys" > "Add new key" on top right (green button). Paste in the new public SSH key we just generated.

# You may also want to go to your local computer shell .ssh folder (same place).
Generate a key and paste into github as done above. This way your local terminal
(not fiji server) can push to github too. Basically any machine(s) you want to
talk to github need this "handshake" SSH key.

Cool we are all connected !


# Setting up direct acess to fiji (no password needed anymore!)

We will be adding the public key to the `~/.ssh/authorized_keys` file.

You'll create a key on your local computer. In windows you can use
puttygen to create the key and on Mac or Linux you can use ssh-keygen
the same way.

The instructions for creating a key with putty are here and detailed.
https://www.ssh.com/ssh/putty/windows/puttygen

The location of the key file should be in cd ~/.ssh for Mac and
in Windows it should be in C:/Users/<username>/.ssh

Then as above for github, you'll copy the public key to your clipboard
and paste it into `~/.ssh/authorized_keys`

```
nano ~/.ssh/authorized_keys
# Then paste in the key and save file.
```
This will allow you to log-in to fiji without having to type your
password in each time.


------------------Nice all connected to Git ----------------------

Now let's see how this all it works:

Navigate to /scratch/Shares/rinnclass/<your name>

# Let's clone the class git repository from GitHub

On the CLASS_2021 GitHub site, click the Code dropdown, select SSH
as your method on connecting and then click the clipboard icon which will
copy the url to your clipboard. You can then paste it here.

```{bash}

git clone <paste-ssh-link>

```
Voila you are connected to our CLASS_2021 git repository.

While in class, I'll make a change to the git and then have you update
your local repository with those changes
using the git pull command.

```{bash}

git pull

```

Note that multiple people could make changes to the same file which results in conflicting versions.

Nicely, git will not allow any commits to be pushed to the remote without asking you to resolve the conflicts between the two versions of a file.

You can minimize the likelihood of having a file conflict if you `git pull` at the
start of each session.


## Create a new branch and make some changes.

You can have as many branches as you want, so in order to avoid conflicts in this case
we'll give you each a branch! Using `git checkout -b` allows you to switch branches,
but it will also create a branch with that name if one doesn't exist.

```{bash}

git checkout -b <your-name>

```

Now use git status to see which branch you're on

```{GIT BASH}

git status

```

## Make some edits and make your first commit

Let's create a new file and type whatever, then you can  `git add`

```{GIT BASH}

nano test.Rmd

# Type in something
# Save file

git status
git add test.Rmd
git status

```

Do you see a new item (test.Rmd) that git noticed was made and ready to sync?

Now we need to commit these changes that have been staged

```{GIT BASH}

git commit -m "Created a new test file."

```

Typically you'll want to be very specific about what you write in the git commit.

Now you should be on the students branch of the git site. Sometimes this can get
buggy.

# Let's try forcing a branch move:

```{GIT BASH}

git push -u origin NEW-BRANCH-NAME

```

Or you may receive an error something like:
fatal: The current branch students has no upstream branch.
To push the current branch and set the remote as upstream, use

    git push --set-upstream origin students

Just paste in that and it will work.

Ok let's update the github website from the terminal!

```{GIT BASH}

git status
git add test.Rmd
git commit -m "(put the notes you want on what was done inside the quotes)"
git push
# type git push again and you should see "up to date"

# !! Always remember to PUSH before you END a session !!
# !! Always remember to PULL before you START a session !!
```

# check out previous changes:

```{GIT BASH}

git log
git diff Xsum1 Xsum2

# if you run into trouble you can always delete a commit.
# For example if you delete your .gitignore file (see below)
# hundreds of files will be staged to be committed.

# to delete a given commit follow these commands:

git log
# find the "key" for the commit you want to delete.
git reset --hard "key"

```

This will show you all the changes -- you can also do this a bit more user friendly on git website.

Finally we need a text file titled ".gitignore" put all the file types you don't want to track. These are large files such as .csv or .pdf etc here are some examples


```{BASH}

touch .gitignore
nano .gitignore

# add extensions of files that are large (pdf, images etc)
# for this class copy and paste the following into .gitignore using nano
```

# PASTE THIS INTO GIT IGNORE file

.Rproj.user
.Rhistory
.RData
.Ruserdata
*.bam
*.pdf
*.Rproj
*.Rproj
*.pdf
*.png
*results*
*figures*
*bam*
*.tmp
*html
*.bed
*_rslurm*
*.chain
*.bed*
*data*
*.txt
*.csv
*archive*
*.rds




-----------------------------------------------
Let's all get set up on fiji and github website using this info ....
