==> class_exercises/09.rmd <==
---
title: "Iranges_GRanges_basics"
author: "JR"
date: "10/6/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(echo = TRUE)
library(GenomicRanges)
library(tidyverse)
# library(Gviz)
library(IRanges)

```

Goal: to understand the basic usage of Iranges and Granges. These are core to almost
all analyses we will perform so it's worth taking sometime to review the fundementals.

? Iranges
? GRanges

These two packages both keep track and can be operated on are: 
start
end
width

You really only need two of these to infer the third. In our case the start and stop
could be a long list of the start and stop of all ChiP-peaks for a given DBP.

Let's make a list of ranges and see what this looks like.
* Note plotRanges is a common function to add to plotting_functions.R

```{r our first range}

# Let's use an example of a ChiP-peak-files each with three peaks. We can use
# this simple example to construct and intersect ranges -- a common procedure

ir1 <- IRanges(start = c(1,200, 1000), end = c(30, 299, 1200))

# take a look
ir1@

# we can use a lot of Iranges functions to look at what we just made
start(ir1)
end(ir1)
width(ir1)

# Cool let's make a second set of peak ranges. We can think of it 
# as a replicate of a DBP with 3 peaks in slightly different position.

ir2 <- IRanges(start = c(17,100, 1100), end = c(49, 199, 1201))

# take a look
ir2 

# nice, so let's concatonate these into one range
ir3 <- c(ir1, ir2)

# take a look -- stacked on top
ir3

# The nice thing about IRanges as we can track values associated with each range
# for example we may want to name these replicate 1 and replicate 2
# let's give it a try

names(ir1) <- paste("replicate_1", 1:3, sep = " ")
ir1

# Nice we now have a new "level" to index that is the name of the range
# or @NAMES in environment. 


# Let's do the same for ir2.
names(ir2) <- paste("replicate_2", 1:3, sep = " ")
ir2
# now concatenate  
ir3 <- c(ir1, ir2)

# let's see
ir3

# nice we have our 'data set' nicely labeled and organized.

```
If you want to plot these you can look here for good code
Not really worth it since you would never plots these typically
https://stackoverflow.com/questions/21506724/how-to-plot-overlapping-ranges-with-ggplot2


# finding overlaps in iranges
```{R overlaps in Iranges}

# first let's use union to "merge" peaks.
# Remaking range 1 since we changed above

ir1 <- IRanges(start = c(1,200, 1000), end = c(30, 299, 1200))
ir2 <- IRanges(start = c(17,100, 1100), end = c(49, 199, 1201))

# union (Iranges function) 
# note: forces the union even if not overlapping!
ov1 <- union(ir1, ir2)

# let's look with ir3 to reference
ir3
ov1
# not exactly what we want but useful let's use intersect:

# intersect (min overlap)
# will only work on overlapping ranges

ov3 <- intersect(ir1, ir2)

# let's look
ir3
ov3
# cool, so this is typically what we want for the min overlap

# we can also look a this with countOveralps to get more info
countOverlaps(ir1, ir2)
# order matters -- but can be figured out with just one countOverlaps
countOverlaps(ir2, ir1)

```

Iranges is great if there is only one chromosome in the genome :) 
However, there are multiple chromosomes in the genome and there are two strands
of DNA that we need to keep track of.

The GenomicRanges package was developed to include the addition information of chr1 and other aspects associated with a given interval in IRanges. Essentially GenomicRanges is a genome specific version of IRanges. 


# Genomic Ranges (Granges)
```{R GRanges}

# Let's create a GRange using IRange logic. What we need is:
# Chromosome, Strand, Ranges

gr <- GRanges(seqnames = c("chrX"), strand = c("+", "-", "*"), ranges = IRanges(start = c(1,200, 1000), end = c(30, 299, 1200)))

# let's see 
gr

# the first thing to note is we see a unique identifier for the X chromosome as
# 'seqnames'(similar to names above)
# 'ranges' (same as Iranges)
# 'strand' 

# we can see all the information with seqinfo
seqinfo(gr)

# we can see some additional place holders we have not changed such as the 
# genome version or if the chromosome is circular. Seqlenght is the lenght of 
# the chromosome. let's add this stuff !

seqlengths(gr) <- c("chrX" = 100000)
seqinfo(gr)

# accessing seq lengths:
seqlengths(gr)

# we see that we only have one 'level' here for ChrX, let's add another chromosome
seqlevels(gr) <- c("chrX", "chrY")
seqlevels(gr)
# now we can see that we have a level that contains both the X nad Y chromosome.


# let's add genome:
genome(gr) <- c("hg38")
gr
# now we have a genome, chromosome and strand associated with each feature!

```


Now let's read in two peak files and find their overlaps!
I have some set up to test in the data folder.

# finding overlapping peaks in our actual data sets !!
```{R find overlaps}

# first we read the consensus peak files in as gRanges object with rtracklayer function.

peaks1 <- rtracklayer::import("/scratch/Shares/rinnclass/CLASS_2022/data/POLR2A_consensus_peaks_filter.bed")

peaks2 <- rtracklayer::import("/scratch/Shares/rinnclass/CLASS_2022/data/POLR2B_consensus_peaks_filter.bed")
 
# now we can use the findOverlaps function to see what overlaps!
?findOverlaps
?import
ovf <- findOverlaps(peaks1, peaks2)
ovf

```

for findOverlaps it is important to remember: 
the first file is query and second is the subject. 
# Thus returns how many peak2 overlaps there were in peaks1 as written above.

# Let's take a look at the overlaps results.
```{r digging into findOverlap results}

summary(ovf)

# Huh 11,000 overlaps -- let's see what we stared with

summary(peaks1)
# Only 10,000 peaks

summary(peaks2)
# 19,000 peaks

# How did we get more than 10,255 peak overlaps (min number of peaks in peaks1)?

# let's see with iranges "width function"
summary(width(peaks1))
summary(width(peaks2))

# there peaks in peak1 are generally larger than peak2
# thus a peak in peak1 could overlap multiple peaks in peak2 (vice versa too).

```

We can get more information from then results of findOverlaps. Let's look at envrionment variable. Note:
@from : peaks from query file indexed into subject file
@to: peaks from subject file indexed into query file.

@from ---------       -----  -----

@to  * * *       *                    *

@from = 1 1 1
@to = 1 2 3 

```{r digging into findOverlap results}

# let's understand this more deeply there are some subtilties we should cover
# let's table the @from index

table(ovf@from) 
table(table(ovf@from))
table(table(table(ovf@from)))
# bit too much info -- let's table the table

table(table(table(table(table(ovf@from)))))

# NOTE: we see how many peaks in peak1 file had multiple overlaps
# thus there is one peak in peak1 that overlaps 14 peaks in peak2

# same for @to
table(table(ovf@to))

# huh most only have one overlap probably because:
# the multiple overlaps in peak2 are assigned to 1 peak in peak1
# 2 peaks in peak2 had overlap with 23 peaks in peak1

```

We can look at how overlaps are reported using: countOverlaps
# countOverlaps

```{r countOverlaps for GRanges}

# let's try countOverlaps
ovf2 <-countOverlaps(peaks1, peaks2)

# let's look
table(ovf2)

# same results as "table(table())

```


*********************
EXCERCISE
*********************


Pick two DBPs in the peaks directory
/scratch/Shares/rinnclass/data/peaks

Each DBP has two replicates -- find the overlaps between replicates.

Then find the percentage of overlaps in each replicate 

If you want to take it a step furhter find the overlaps between the overlaps in replicates.

Or in other words the overlap between two differnt DBPs "consensus peaks"
