==> class_exercises/14_2.rmd <==
---
title: "our_data_in_ranges"
author: "JR"
date: "11/3/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(GenomicRanges)
# source("../../../util/intersect_functions.R")
source("../../../util/plotting_functions.R")
source("../../../util/_setup.R")
library(Gviz)
library(ggpubr)
library(ggplot2)
```


Alright time to start digging into the data for new results !!
We have made a consensus peak file for each DBP and 
Moreover, we have a data frame of a bunch of peak overlap features for each DBP.

# Loading in filtered consensus peaks

First let's load in the consensus peak files (this is something we always have to do)

```{r loading consensus peaks}

# read in file list to lapply import below

# first create a file list of filtered consensus peaks
fl <- list.files("/scratch/Shares/rinnclass/CLASS_2022/data/filtered_consensus_peaks", 
                                             pattern = "*.bed",
                                             full.names = TRUE)


# # lappy to import each file in file list 
filtered_consensus_peaks <- lapply(fl, rtracklayer::import)

# fl <- list.files("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/class_exeRcises/analysis/11_consensus_peaks/filtered_consensus_peaks/", full.names=TRUE)
# 
# # loading each consensus peak bed file into a list of GRanges
# filtered_consensus_peaks <- lapply(fl, rtracklayer::import, format = "bed")


# Let's take a look since this is an important first step.
# Ok we should probably add some names -- another common task.
# we can get name from @elementdata$listdata$name
filtered_consensus_peaks[[40]]$name %>%
  head()

# cool now we can strsplit this to get the name 
# on the left of  x_ [1] indicates take the left
# let's use strplit to clean the name up by splitting on _

?strsplit
# let's add names to our consensus peaks list of GRanges
# selecting the left of the split with index [1]


names(filtered_consensus_peaks) <- sapply(filtered_consensus_peaks, function(x){
  unlist(strsplit(x$name, "_"))[[1]]
})

# Now we have a named list of consensus peaks in the environment -- let's take a look
```

# loading in num_peaks_df
Let's load in our num_peaks_df and decide if we want to filter?
```{r num_peaks_df}

# loading in our previous analyses:

num_peaks_df <- read.csv("/scratch/Shares/rinnclass/CLASS_2022/<your_folder>/CLASS_2022/class_exeRcises/analysis/12_peak_features/num_peaks_df.csv")

# num_peaks_df <- read.csv("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/class_exeRcises/analysis/12_peak_features/num_peaks_df.csv")


```

# Let's plot some data and see what we might learn
Ok enough set up let's dig in with making some plots!


```{r plotting peak features}

# First let's look at a histogram of peak#/DBP

 ggplot(num_peaks_df, aes(x = num_peaks)) + 
  geom_histogram(bins = 70)


```


# plotting num_peaks versus genome coverage.
? What would you suspect would happen
```{r peaks vs coverage}

# Let's plot our result of num_peaks versus genome coverage.
ggplot(num_peaks_df, aes(x = num_peaks, y = total_peak_length)) +
  geom_point() + 
  geom_smooth(method = "gam", se = TRUE, color = "black", lty = 2)+
              #formula = 'y ~ x') +
  # after this line is just formatting figure - above is "functional"
  #geom_text() +
  ylab("BP covered") +
  xlab("Number of peaks") +
  ggtitle("Peak count vs. total bases covered")

# saving MAKE SURE WORKING DIRECTORY IS CORRECT
ggsave("figures/peak_num_vs_coverage.pdf")

# Feel free to copy code from 14_plotting_R_part_I.Rmd for other versions of this figure.

```

# plotting num peaks on promoters
Let's did in a bit deeper and see how coverage on promoters.
```{r number of DBPS on promoters}
# now let's plot x as num_peaks and y num overlapping promoters.

ggplot(num_peaks_df,
       aes(x = num_peaks, y = peaks_overlapping_promoters)) +
  # just a resource for figure editing
  xlab("Peaks per DBP") +
  ylab("Number of peaks overlapping promoters") +
  ggtitle("Relationship Between Number of DBP Peaks and Promoter Overlaps")+
  geom_point() +
  
  # back to important statistical aspects
  geom_abline(slope = 1, linetype="dashed") +
  geom_smooth(method = "lm", se=FALSE, formula = 'y ~ x',
              color = "#a8404c") +
  # adding a seconf
  stat_regline_equation(label.x = 35000, label.y = 18000) +
  ylim(0,60100) +
  xlim(0,60100)

ggsave("figures/3_peak_num_vs_promoter_coverage.pdf")

# we observe a "saturation of binding events -- as you get more
# peaks you stop increaseing binding to promoters.
# maybe it leaks into the gene body let's check
```


# peak Coverage on gene bodies
```{r peak coverage on gene bodies}

ggplot(num_peaks_df,
       aes(x = num_peaks, y = peaks_overlapping_genebody)) +
  
# code to make things pretty
  xlab("Peaks per DBP") +
  ylab("Number of peaks overlapping genes") +
  ggtitle("Relationship Between Number of DBP Peaks and Gene Body Overlaps")+
  geom_point() +
  geom_abline(slope = 1, linetype="dashed") +
  geom_smooth(method = "lm", se=F, formula = 'y ~ x',
              color = "#a8404c") +
  stat_regline_equation(label.x = 35000, label.y = 18000) +
  ylim(0,60100) +
  xlim(0,60100)

# Interesting result !! Gene bodies explain almost all the places of binding in 
# the genome! Where as promoters had a non linear asymptope after a certain number 
# of peaks. VERY COOL and surprising. There are billions of other possible binding sites!

ggsave("/figures/4_peak_num_vs_gene_body_coverage.pdf")
```

Ok so we have started generating results in this handy num_peaks_df. 
However, the peak_occurence_df is also very handy for analyses.
Let's load it in and do some analyses -- THIS ONE IS REALLY COOL :).

# Reading in Peak occurence df

```{r read in peak_occurence_df}

# First read the file into an object

peak_occurence_df <- read.csv("/scratch/Shares/rinnclass/CLASS_2022/<your_folder>/CLASS_2022/class_exeRcises/analysis/12_peak_features/peak_occurence_dataframe.csv")

# peak_occurence_df <- read.csv("/scratch/Shares/rinnclass/CLASS_2022/JR/CLASS_2022/class_exeRcises/analysis/12_peak_features/peak_occurence_dataframe.csv")

# let's have a look. This is a summary of the peak occurence df.
# we did colsums to find how many binding events there are at each promoter
# so the last column will tell us how many binding events there were at a given promoter. 

```

Let's make a density plot of num DBPs bound per promoter
# Density plot of binding events

```{r density plot of DBP localization events}

# Since we are making a density plot we only need the X axis.

ggplot(peak_occurence_df, aes(x = number_of_dbp)) +
geom_density(alpha = 0.2, color = "#424242", fill = "#424242") +
  
  theme_paperwhite() +
  xlab(expression("Number of DBPs")) +
  ylab(expression("Density")) +
  ggtitle("Promoter binding events",
          subtitle = "mRNA and lncRNA genes") 

# Wow what a result! Last year we only saw up to 111 binding events at a promoter
# this is kind of validated where the first peak in the distribution falls at 100
# binding events -- yet there is a second distribution of regions of 300 or more 
# binding events -- wonder what those are :) ?

# VERY COOL let's print out the figure

ggsave("figures/num_binding_events_per_promoter.pdf")

```


Lets find how many promoters don't have any DBPs bound
# promoters with out binding events

```{r prmoters with out binding events}

# here is just a simple index and filter of the index to have at least 1 dbp bound.
unbound_promoters <- peak_occurence_df %>% 
  filter(peak_occurence_df$number_of_dbp < 1)

# how many are there?
nrow(unbound_promoters)
# so there are only a few 6,720 promoters that don't have binding evetns (~10%)

#  let's put it in a folder called results. We will always use this folder structure
write_csv(unbound_promoters, "results/unbound_promoters.csv")

```


# Do any DBPs never bind a promoter?
```{r is there a DBP that doesn't bind promoters?}

# we actually don't need code for this we can look in num_peaks_df and sort :)
# There are only a few and we may want to filter these out above.

```

# lncRNA versus mRNA promoter binding
Let's compare the binding patterns of lncRNA vs mRNA promoters.
While we are at it let's do some deep dives into the aes layer
This is required to make two plots in one :)

```{r}

# Plotting lncRNA -vs mRNA promoter binding distribution
# for more advanced figure making check out the code below.
# it is not the best practice to do to many analyses in ggplot.
# The below esxample shows how powerful and flexible ggplot is.

ggplot(num_peaks_df, aes(x = num_peaks)) +
  geom_point(aes(y = peaks_overlapping_lncrna_promoters), color = "red") +
  geom_point(aes(y = peaks_overlapping_mrna_promoters), color = "black") +
  stat_regline_equation(aes(y = peaks_overlapping_lncrna_promoters), color = "red") +
  stat_regline_equation(aes(y = peaks_overlapping_mrna_promoters), color = "black", label.y = 20000) +
  geom_smooth(aes(y = peaks_overlapping_lncrna_promoters), method = "lm", se = FALSE, formula = "y ~ x") +
  geom_smooth(aes(y = peaks_overlapping_mrna_promoters), method = "lm", se = FALSE, formula = "y ~ x")

ggsave("figures/lncRNA-mRNA-promoter_binding.pdf")

# Nice we see that mRNA promoters have about 4x more binding (from slope)
# Both follow a fairly linear trend except mRNA may have saturate quicker.


# This is just a fancier version of the same thing -- note mutate to clean up names
# this first part is manipulating the num peaks df with pivot longer
num_peaks_dfl <- num_peaks_df %>%
  dplyr::select(-peaks_overlapping_promoters) %>%
  # selecting these to cols to pivot longer
  # We will discuss pivot longer in the next class (BUT IF YOUR INTERESTED ASK NOW IT'S WORTH IT!!)
  pivot_longer(cols = peaks_overlapping_lncrna_promoters:peaks_overlapping_mrna_promoters,
               names_to = "gene_type",
               values_to = "peaks_overlapping_promoters") %>%
  mutate(gene_type = gsub("peaks_overlapping_", "", gene_type))

# plotting
ggplot(num_peaks_dfl, aes(x = num_peaks, y = peaks_overlapping_promoters, 
                         col = gene_type)) +
         geom_point() +
         geom_abline(slope = 1, linetype="dashed") +
  geom_smooth(method = "lm", se = FALSE, formula = "y ~ x") +
  stat_regline_equation() +
  scale_color_manual(values = c("#a8404c", "#424242"))+
  xlab("Peaks per DBP") +
  ylab("Peaks Overlapping Promoters") +
  ggtitle("Number of DBP Peaks and Promoter Overlaps")
#ggsave("figures/peaks_overlaps_relationship_by_gene_type.png", height = 5, width = 8)
ggsave("figures/peaks_overlaps_relationship_by_gene_type.pdf", height = 5, width = 8)
``` 

############
exeRcise:
############

Make a figure of percentage of peaks for each DBP that overlap promoters & gene bodies.
Add these columns to num_peaks_df
